Wip add support for generics make `as` authoritative
This commit further enhances the implementation for full generic
support. From now on the full type paths need used to declare the types
in the `OpenApi` and `#[utoipa::path]` macros.

As further elaborated here https://github.com/juhaku/utoipa/pull/1020#issue-2503474650
the name of the schema is now resolved form the type name and possible
generic arguments or via `as = ...` attribute and possible generic
arguments. This name is then used across the OpenApi. This makes it a
single place to define the name or prefixed name for the type unlike
previously the path of the schema in `OpenApi` macro or `request_body`
or `response body` was added to the name.

This is a breaking change.

Fixes #993 Closes #1020